Adds daemon and translation script
This commit adds the `gtd` implementation which is looking for the data
file and the target language in the designated queue in order to start
the actual translation of the text. For each sentence in the data file
it starts a new process and in the end, it waits for all processes to
finish before sending the actual translation of the text to gtranslate
script using the same queue as the `gtranslate`. On the other hand,
`gtranslate` is expecting from the user as an input the name of the
file to read the sentences and the target language to do the translation.
Once the `gtd` sends the translation back to `gtranslate`, the second one
will display the data to the user.